+++ src/react-native/xcode.ts
@@ -145,46 +145,6 @@ REACT_NATIVE_XCODE="$REACT_NATIVE_PATH/scripts/react-native-xcode.sh"
   return patchedScript;
 }
 
-export function addSentryWithCliToBundleShellScript(
-  script: string,
-): string | ErrorPatchSnippet {
-  if (!script.includes('$REACT_NATIVE_XCODE')) {
-    clack.log.error(
-      `Could not find $REACT_NATIVE_XCODE in ${chalk.cyan(
-        'Bundle React Native code and images',
-      )} build phase. Skipping patching.`,
-    );
-    Sentry.captureException(
-      `Could not find $REACT_NATIVE_XCODE in 'Bundle React Native code and images' build phase.`,
-    );
-    return new ErrorPatchSnippet(
-      makeCodeSnippet(true, (unchanged, plus, _minus) => {
-        return unchanged(`${plus(`export SENTRY_PROPERTIES=sentry.properties
-export EXTRA_PACKAGER_ARGS="--sourcemap-output $DERIVED_FILE_DIR/main.jsbundle.map"
-`)}
-/bin/sh -c "$WITH_ENVIRONMENT ${plus(
-          `\\"../node_modules/@sentry/cli/bin/sentry-cli react-native xcode`,
-        )} $REACT_NATIVE_XCODE${plus(`\\"`)}"
-${plus(
-  `/bin/sh -c "$WITH_ENVIRONMENT ../node_modules/@sentry/react-native/scripts/collect-modules.sh"`,
-)}
-`);
-      }),
-    );
-  }
-  return (
-    'export SENTRY_PROPERTIES=sentry.properties\n' +
-    'export EXTRA_PACKAGER_ARGS="--sourcemap-output $DERIVED_FILE_DIR/main.jsbundle.map"\n' +
-    script.replace(
-      '$REACT_NATIVE_XCODE',
-      () =>
-        // eslint-disable-next-line no-useless-escape
-        '\\"../node_modules/@sentry/cli/bin/sentry-cli react-native xcode $REACT_NATIVE_XCODE\\"',
-    ) +
-    '\n/bin/sh -c "$WITH_ENVIRONMENT ../node_modules/@sentry/react-native/scripts/collect-modules.sh"\n'
-  );
-}
-
 export function addDebugFilesUploadPhaseWithBundledScripts(
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   xcodeProject: any,
@@ -207,43 +167,6 @@ export function addDebugFilesUploadPhaseWithBundledScripts(
     {
       shellPath: '/bin/sh',
       shellScript: `/bin/sh ../node_modules/@sentry/react-native/scripts/sentry-xcode-debug-files.sh`,
-    },
-  );
-  clack.log.success(
-    `Added Build phase ${chalk.cyan('Upload Debug Symbols to Sentry')}.`,
-  );
-}
-
-export function addDebugFilesUploadPhaseWithCli(
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  xcodeProject: any,
-  { debugFilesUploadPhaseExists }: { debugFilesUploadPhaseExists: boolean },
-) {
-  if (debugFilesUploadPhaseExists) {
-    clack.log.warn(
-      `Build phase ${chalk.cyan(
-        'Upload Debug Symbols to Sentry',
-      )} already exists.`,
-    );
-    return;
-  }
-
-  xcodeProject.addBuildPhase(
-    [],
-    'PBXShellScriptBuildPhase',
-    'Upload Debug Symbols to Sentry',
-    null,
-    {
-      shellPath: '/bin/sh',
-      shellScript: `
-WITH_ENVIRONMENT="../node_modules/react-native/scripts/xcode/with-environment.sh"
-if [ -f "$WITH_ENVIRONMENT" ]; then
-  . "$WITH_ENVIRONMENT"
-fi
-export SENTRY_PROPERTIES=sentry.properties
-[ "$SENTRY_INCLUDE_NATIVE_SOURCES" = "true" ] && INCLUDE_SOURCES_FLAG="--include-sources" || INCLUDE_SOURCES_FLAG=""
-../node_modules/@sentry/cli/bin/sentry-cli debug-files upload "$INCLUDE_SOURCES_FLAG" "$DWARF_DSYM_FOLDER_PATH"
-`,
     },
   );
   clack.log.success(